remove version numbers from column names
Pandas adds numbers to duplicate column names (‘.1’, ‘.2’, etc).
This is a problem when plotting replicates from different columns
because they get treated as independent samples
The solution is to remove these numbers before appending to the traces,
but names may have periods in them on purpose.
The best I could come up with is to make sure that the characters after
the last period are integers and to issue a warning if they are removed
It looks a bit goofy to do this with a try/except statement but it
seems to work